+++ R/p_ceiling.R
@@ -29,12 +29,12 @@ function (loop.data, slope, intercept) {
   }
 
   if (slope > 0) {
-    # Line doesn't go through the left boundary but bottom boundry
+    # Line doesn't go through the left boundary but lower boundry
     if (p.l[2] < theo[3]) {
       p.l <- c((theo[3] - intercept) / slope, theo[3])
     }
 
-    # Line doesn't go through the right boundary but top boundry
+    # Line doesn't go through the right boundary but upper boundry
     if (p.r[2] > theo[4]) {
       p.r <- c((theo[4] - intercept) / slope, theo[4])
     }
@@ -44,12 +44,12 @@ function (loop.data, slope, intercept) {
     area_above <- area_above + (p.l[1] - theo[1]) * (theo[4] - theo[3])
     area_above <- area_above - (p.l[1] - theo[1]) * (theo[4] - p.r[2])
   } else if (slope < 0) {
-    # Nope, line goes through top boundary
+    # Nope, line goes through upper boundary
     if (p.l[2] > theo[4]) {
       p.l <- c((theo[4] - intercept) / slope, theo[4])
     }
 
-    # Line doesn't go through bottom boundary but right boundary
+    # Line doesn't go through lower boundary but right boundary
     if (p.r[2] < theo[3]) {
       p.r <- c((theo[3] - intercept) / slope, theo[3])
     }
@@ -76,7 +76,7 @@ function (loop.data, peers, method) {
   ceiling <- p_scope_ceiling(peers, theo, emp, flip.x, flip.y)
 
   # Single peer, different cases for FHD+VRS and conf lines
-  if (is.vector(peers) || length(peers) <= 2) {
+  if (nrow(unique(peers)) <= 1) {
     # if there is only one peer, the ceiling zone is zero or 'scope ceiling'
     if (method %in% c("fdh", "vrs")) {
       return ( ceiling )
@@ -123,25 +123,25 @@ function (peers, theo, emp, flip.x, flip.y) {
 
   left    <- x1 * (theo[4] - theo[3])
   right   <- x2 * (theo[4] - theo[3])
-  bottom  <- (theo[2] - theo[1]) * y1
-  top     <- (theo[2] - theo[1]) * y2
+  lower   <- (theo[2] - theo[1]) * y1
+  upper   <- (theo[2] - theo[1]) * y2
 
   if (!flip.x && !flip.y) {
     # Upper left corner
     cross <- x1 * y2
-    return (left + top - cross)
+    return (left + upper - cross)
   } else if (flip.x && !flip.y) {
     # Upper right corner
     cross <- x2 * y2
-    return (right + top - cross)
+    return (right + upper - cross)
   } else if (flip.x && flip.y) {
     # Lower right corner
     cross <- x2 * y1
-    return (right + bottom - cross)
+    return (right + lower - cross)
   } else if (!flip.x && flip.y) {
     # Lower left corner
     cross <- x1 * y1
-    return (left + bottom - cross)
+    return (left + lower - cross)
   }
 }