+++ R/p_nca_ce_fdhi.R
@@ -4,7 +4,8 @@ function (loop.data, bn.data) {
   flip.y <- loop.data$flip.y
 
   # Find the points on the ceiling ("PEERS")
-  peers       <- p_peers(loop.data, inflate=TRUE, limit=TRUE)
+  peers       <- p_peers(loop.data)
+  # peers       <- p_peers(loop.data, inflate=TRUE, limit=TRUE)
 
   line        <- p_fdh_line(loop.data$scope.theo, peers, flip.x, flip.y)
   ceiling     <- p_ce_ceiling (loop.data, peers, "fdh")
@@ -13,7 +14,7 @@ function (loop.data, bn.data) {
   bottleneck  <- p_bottleneck_ce(loop.data, bn.data, peers, "fdh")
   fit         <- get_fit(ceiling, loop.data$ce_fdh_ceiling)
 
-  return(list(line=line,
+  return(list(line=line, peers=peers,
               slope=NA, intercept=NA,
               ceiling=ceiling, effect=effect,
               above=0, accuracy=100, fit=fit,