+++ R/p_nca_cr_cm.R
@@ -5,15 +5,15 @@ function (loop.data, bn.data) {
   weighting <- loop.data$weighting
 
   columns     <- p_columns(loop.data, FALSE)
-  cm_peers    <- t(columns[c(2, 5), ])
+  peers    <- t(columns[c(2, 5), ])
 
-  if (!is.vector(cm_peers) && length(cm_peers) > 2) {
+  if (nrow(unique(peers)) > 1) {
     # Get the weights if required
     weights <- if (weighting) columns[1,]
 
     # Perform OLS through the peers
-    x <- cm_peers[,1]
-    y <- cm_peers[,2]
+    x <- peers[, 1]
+    y <- peers[, 2]
     line <- lm(y~x, weights=weights)
     
     intercept <- unname(coef(line)["(Intercept)"])
@@ -26,6 +26,7 @@ function (loop.data, bn.data) {
     intercept <- NA
     slope     <- NA
     above     <- 0
+    peers     <- matrix(, nrow = 0, ncol = 2)
   }
 
   effect      <- ceiling / loop.data$scope.area
@@ -34,7 +35,7 @@ function (loop.data, bn.data) {
   ineffs      <- p_ineffs(loop.data, slope, intercept)
   bottleneck  <- p_bottleneck(loop.data, bn.data, slope, intercept)
 
-  return(list(line=line,
+  return(list(line=line, peers=peers,
               slope=slope, intercept=intercept,
               ceiling=ceiling, effect=effect,
               above=above, accuracy=accuracy, fit=fit,