+++ R/p_nca_cr_fdhi.R
@@ -5,9 +5,10 @@ function (loop.data, bn.data) {
   weighting <- loop.data$weighting
 
   # Find the points on the ceiling ("PEERS")
-  peers <- p_peers(loop.data, inflate=TRUE)
+  peers <- p_peers(loop.data)
+  # peers <- p_peers(loop.data, inflate=TRUE)
 
-  if (!is.vector(peers) && length(peers) > 2) {
+  if (nrow(unique(peers)) > 1) {
     # Get the weights if required, from the non-inflated peers
     weights <- if (weighting) p_weights(loop.data, p_peers(loop.data))
 
@@ -26,6 +27,7 @@ function (loop.data, bn.data) {
     intercept <- NA
     slope     <- NA
     above     <- 0
+    peers     <- matrix(, nrow = 0, ncol = 2)
   }
 
   effect      <- ceiling / loop.data$scope.area
@@ -34,7 +36,7 @@ function (loop.data, bn.data) {
   ineffs      <- p_ineffs(loop.data, slope, intercept)
   bottleneck  <- p_bottleneck(loop.data, bn.data, slope, intercept)
 
-  return(list(line=line,
+  return(list(line=line, peers=peers,
               slope=slope, intercept=intercept,
               ceiling=ceiling, effect=effect,
               above=above, accuracy=accuracy, fit=fit,